Modify Razor_AHRS to work with IGVC
The Razor_AHRS has been modified to publish a ros message which
consists of 3 floats for roll, pitch and yaw (rpy).  These will be
converted upstream into ros conforming quaternions.

The code had to be modified slightly to accomodate fitting on an
arduino uno and so rpy is only held in the ros message and some of the
error correction and all of the console output has been removed.